Refactored: Formatted and file name changes
