+++ frontend/src/components/common/Table_2.jsx
