+++ frontend/src/components/common/notifications/ProgressCard.jsx
